+++ d912pxy/d912pxy_helper.h
@@ -63,9 +63,9 @@ namespace d912pxy_helper {
 	char* StrGetCurrentLineStart(char* buffer);
 	char* StrNextLine(char* buffer);
 
-	UINT64 GetClosestPow2(UINT64 size);
+	UINT GetClosestPow2(UINT size);
 
-	UINT64 AlignValueByPow2(UINT64 val, UINT64 pow2val);
+	UINT AlignValueByPow2(UINT val, UINT pow2val);
 
 	d912pxy_file_path* GetFilePath(d912pxy_file_path_id fpId);
