+++ d912pxy/d912pxy_upload_pool.h
@@ -32,13 +32,13 @@ public:
 	d912pxy_upload_item(UINT8 icat);
 	~d912pxy_upload_item();
 
-	void UploadBlockWrite(UINT64 dst_offset, UINT64 upload_offset, UINT64 sz, void* src);
-	void UploadBlock(ID3D12Resource * res, UINT64 dst_offset, UINT64 upload_offset, UINT64 sz, void* src, ID3D12GraphicsCommandList* cl);
+	void UploadBlockWrite(UINT64 dst_offset, UINT64 upload_offset, UINT sz, void* src);
+	void UploadBlock(ID3D12Resource * res, UINT64 dst_offset, UINT64 upload_offset, UINT sz, void* src, ID3D12GraphicsCommandList* cl);
 
-	void UploadTargetWithOffset(ID3D12Resource * res, UINT64 sofs, UINT64 dofs, UINT64 sz, void* src, ID3D12GraphicsCommandList* cl);
-	void UploadTarget(ID3D12Resource* res, UINT64 dofs, UINT64 sz, void* src, ID3D12GraphicsCommandList* cl);
+	void UploadTargetWithOffset(ID3D12Resource * res, UINT sofs, UINT dofs, UINT sz, void* src, ID3D12GraphicsCommandList* cl);
+	void UploadTarget(ID3D12Resource* res, UINT dofs, UINT sz, void* src, ID3D12GraphicsCommandList* cl);
 	
-	void Reconstruct(void* mem, UINT64 rowPitch, UINT64 height, UINT64 size, UINT64 upload_offset, const D3D12_RANGE* wofs);
+	void Reconstruct(void* mem, UINT rowPitch, UINT height, UINT size, UINT64 upload_offset, const D3D12_RANGE* wofs);
 
 	ID3D12Resource* GetResourcePtr() { return mRes; };
 
@@ -52,14 +52,14 @@ public:
 
 	intptr_t GetCurrentOffset() { return usedSpace; };
 
-	void AddSpaceUsed(UINT64 amount) { usedSpace += amount; };
+	void AddSpaceUsed(size_t amount) { usedSpace += amount; };
 	   
 private:	
-	intptr_t DPtr();
+	UINT64 DPtr();
 
-	intptr_t DPtrOffset(UINT64 offset);
+	UINT64 DPtrOffset(UINT64 offset);
 
-	intptr_t mappedMemWofs;
+	UINT64 mappedMemWofs;
 
 	intptr_t usedSpace;
 	intptr_t space;
